requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
Authlib==1.3.1
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
email_validator==2.2.0
fastapi==0.115.8
fastapi-cli==0.0.7
frozenlist==1.5.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-aiplatform==1.80.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-health-checking==1.70.0
grpcio-status==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.18
langchain-core==0.3.35
langchain-google-vertexai==2.0.13
langchain-neo4j==0.3.0
langchain-openai==0.3.3
langchain-text-splitters==0.3.6
langsmith==0.3.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
neo4j==5.27.0
numpy
openai==1.61.0
orjson==3.10.15
packaging==24.2
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
rdflib==7.1.3
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
setuptools==75.8.0
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
validators==0.34.0
watchfiles==1.0.4
weaviate-client==4.10.4
websockets==14.2
yarl==1.18.3
zstandard==0.23.0
